requirements.txt
apache-airflow-providers-mongo==1.0.1